+++ app-proto.js
@@ -1,6 +1,6 @@
 /* debug */
 
-const DEBUG = false;
+const DEBUG = true;
 
 /* dependencies */
 
@@ -28,6 +28,15 @@ console.log("Server started.");
 
 var world = {
 	size: 21,
+	layer: {
+		size: 3,
+		indexing: {
+			"bottom": "0",
+			"mid_bottom": undefined,
+			"mid_top": "1",
+			"top": "2",
+		}
+	},
 
 	build(width, height, layer) {
 		for(let l = 0; l < layer; l++) {
@@ -35,7 +44,7 @@ var world = {
 			for(let y = 0; y < height; y++) {
 				this.map[l].push([]);
 				for(let x = 0; x < width; x++) {
-					this.map[l][y].push(0);
+					this.map[l][y].push( 0 );
 				}
 			}
 		}
@@ -49,22 +58,23 @@ var world = {
 			}
 		}
 	},
-
+	
 	make(width, height, layer, socket) {
 		this.map = [];
-
+		
 		this.build(width, height, layer);
 		this.edit(width, height, layer);
-
+		
 		if( socket ) { socket.emit("new_map", this.map) }
 		return this.map;
 	},
-
+	
 	find(l, x, y) {
 		return this.map[l][y + ((this.size - 1) / 2)][x + ((this.size - 1) / 2)];
 	}
 };
 
+world.make(world.size, world.size, world.layer.size);
 
 
 // --------------------------------------------------------------------------------------------------------------------
@@ -89,86 +99,72 @@ class entity {
 var PLAYER_LIST = {};
 
 class player extends entity {
-	constructor(x, y, spd_x = 0, spd_y = 0, img = 1) {
+	constructor(x, y, img = 1) {
 		super(img, x, y);
-
-		this.spd_x = spd_x;
-		this.spd_y = spd_y;
 
 		this.spd = 1;
 
+		this.commands = [];
+
 		this.going = {
-			left  : 0,
-			up    : 0,
-			right : 0,
-			down  : 0,
-		};
-
-		// this.pressing = {
-		// 	left  : false,
-		// 	up    : false,
-		// 	right : false,
-		// 	down  : false
-		// };
+			left  : false,
+			up    : false,
+			right : false,
+			down  : false
+		}
 	}
-
+	
 	update() {
+		super.update();
+
 		this.update_spd();
 		this.update_pos();
-		super.update();
 	}
 
-	update_pos() {
-		this.x += this.spd_x;
-		this.y += this.spd_y;
+	update_cmd(direction) {
+		this.going[direction] -= 1;
+		this.commands.push(direction);
 	}
 
 	update_spd() {
-		// if( this.pressing.left  ) { this.spd_x = -this.spd }
-		// if( this.pressing.up    ) { this.spd_y =  this.spd }
-		// if( this.pressing.right ) { this.spd_x =  this.spd }
-		// if( this.pressing.down  ) { this.spd_y = -this.spd }
-		//
-		// if( !( this.pressing.left || this.pressing.right ) ) { this.spd_x = 0; }
-		// if( !( this.pressing.up   || this.pressing.down  ) ) { this.spd_y = 0; }
+		if(this.going.left  ) { this.update_cmd( "left"  ) }
+		if(this.going.up    ) { this.update_cmd( "up"    ) }
+		if(this.going.right ) { this.update_cmd( "right" ) }
+		if(this.going.down  ) { this.update_cmd( "down"  ) }
+	}
 
-		// if( this.pressing.left  ) { this.spd_x = -this.spd }
-		// if( this.pressing.up    ) { this.spd_y =  this.spd }
-		// if( this.pressing.right ) { this.spd_x =  this.spd }
-		// if( this.pressing.down  ) { this.spd_y = -this.spd }
-		//
-		// if( !( this.pressing.left || this.pressing.right ) ) { this.spd_x = 0; }
-		// if( !( this.pressing.up   || this.pressing.down  ) ) { this.spd_y = 0; }
+	update_pos() {
+		var current = this.commands.shift();
+
+		if( current === "left"  ) { this.x -= 1 }
+		if( current === "up"    ) { this.y += 1 }
+		if( current === "right" ) { this.x += 1 }
+		if( current === "down"  ) { this.y -= 1 }
 	}
 
 	on_connect(socket, id) {
 		PLAYER_LIST[id] = this;
 
 		socket.on("key_press", (data) => {
-			// if( data.input_id === "left"  ) { this.pressing.left  = data.state }
-			// if( data.input_id === "up"    ) { this.pressing.up    = data.state }
-			// if( data.input_id === "right" ) { this.pressing.right = data.state }
-			// if( data.input_id === "down"  ) { this.pressing.down  = data.state }
 			if( data.input_id === "left"  ) { this.going.left  += 1 }
 			if( data.input_id === "up"    ) { this.going.up    += 1 }
 			if( data.input_id === "right" ) { this.going.right += 1 }
 			if( data.input_id === "down"  ) { this.going.down  += 1 }
 		});
-
-		world.make(world.size, world.size, 3);
-
+		
 		socket.emit("connection", {
 			world : {
-				DEBUG : DEBUG      ,
-				map   : world.map  ,
-				size  : world.size ,
-				msg   : `You are now connected!\nYour session id is now: ${id}`
+				DEBUG : DEBUG       ,
+				map   : world.map   ,
+				size  : world.size  ,
+				layer : world.layer ,
+				msg   : `You are now connected!\nYour session id is now: ${id}`,
 			},
 			me    : {
-				id    : id         ,
-				x     : this.x     ,
-				y     : this.y     ,
-				size  : this.size  ,
+				id    : id          ,
+				x     : this.x      ,
+				y     : this.y      ,
+				size  : this.size   ,
 				img   : this.img
 			},
 		});
@@ -198,34 +194,35 @@ io.sockets.on("connection", function(socket) {
 	});
 
 	socket.on("send_msg_to_server", (data) => {
-		if(data.from.id === id && data.msg[0] !== "<<>>") {  // if user has matching credentials
+		if(data.from.id === id && data.msg[0] !== ";") {  // if user has matching credentials
 
 			//  if user has a name & it matches provided name, then broadcast message
 			if(  p.name && data.from.name === p.name ) { socket.broadcast.emit("add_to_chat", { from: {name: p.name}, msg: data.msg }) }
 
 			//  if user has a name & it doesn't match provided name, then tell them they can't do that and refuse to send message
-			if(  p.name && data.from.name !== p.name ) { socket.emit("add_to_chat", { from: {name: "<<>>", id: "<<>>" }, msg: "You can't change your name!", my_name: p.name }); }
+			if(  p.name && data.from.name !== p.name ) { socket.emit("add_to_chat", { from: {name: ";", id: ";" }, msg: "You can't change your name!", my_name: p.name }); }
 
 			//  if name hasn't been set, then set the name and send message
 			if( !p.name  ) {
 				Object.defineProperty(p, "name", { value: data.from.name, writable: false });
-				          socket.emit("add_to_chat", { from: { name: "<<>>"   , id: "<<>>" }, msg: `Your session name is now: ${p.name}`, my_name: p.name });
+				          socket.emit("add_to_chat", { from: { name: ";"   , id: ";" }, msg: `Your session name is now: ${p.name}`, my_name: p.name });
 				socket.broadcast.emit("add_to_chat", { from: { name: p.name          }, msg: data.msg });
 			}
 		}
 
-		if(data.from.id !== p.id) // disconnect, nothing yet
+		if(data.from.id !== p.id) {} // disconnect, nothing yet
 
 		exec_debug(socket, p, data); // run command
 	});
 });
 
 function exec_debug(socket, p, data) {
-	if( DEBUG === true && data.msg[0] === "<<>>" ) {
-		socket.emit("add_to_chat", { from: { name: "<<>>", id: "<<>>" }, msg: "You have issued a command" });
+	if( DEBUG === true && data.msg[0] === ";" ) {
+		socket.emit("add_to_chat", { from: { name: ";", id: ";" }, msg: "You have issued a command" });
 		return eval(data.msg.substr(1));
 	}
 }
+
 function emit_debug(socket, p, data) { socket.emit("debug", data.msg) }
 
 
@@ -243,6 +240,7 @@ function update_pckgs() {
 		pack[player] = {
 			x    : PLAYER_LIST[player].x,
 			y    : PLAYER_LIST[player].y,
+			img  : PLAYER_LIST[player].img,
 			size : PLAYER_LIST[player].size
 		};
 	}
@@ -252,6 +250,6 @@ function update_pckgs() {
 
 setInterval(() => {
 	var pack = update_pckgs();
-
+	
 	io.emit("update", pack);
-}, ( 1000/20 ) ); // prev 1000 / 8
+}, ( 1000/10 ) ); // prev 1000 / 8