+++ app.js
@@ -1,6 +1,6 @@
 /* debug */
 
-const DEBUG = false;
+const DEBUG = true;
 
 /* dependencies */
 
@@ -28,6 +28,15 @@ console.log("Server started.");
 
 var world = {
 	size: 21,
+	layer: {
+		size: 3,
+		indexing: {
+			"bottom": "0",
+			"mid_bottom": undefined,
+			"mid_top": "1",
+			"top": "2",
+		}
+	},
 
 	build(width, height, layer) {
 		for(let l = 0; l < layer; l++) {
@@ -49,22 +58,23 @@ var world = {
 			}
 		}
 	},
-
+	
 	make(width, height, layer, socket) {
 		this.map = [];
-
+		
 		this.build(width, height, layer);
 		this.edit(width, height, layer);
-
+		
 		if( socket ) { socket.emit("new_map", this.map) }
 		return this.map;
 	},
-
+	
 	find(l, x, y) {
 		return this.map[l][y + ((this.size - 1) / 2)][x + ((this.size - 1) / 2)];
 	}
 };
 
+world.make(world.size, world.size, world.layer.size);
 
 
 // --------------------------------------------------------------------------------------------------------------------
@@ -135,21 +145,20 @@ class player extends entity {
 			if( data.input_id === "right" ) { this.pressing.right = data.state }
 			if( data.input_id === "down"  ) { this.pressing.down  = data.state }
 		});
-
-		world.make(world.size, world.size, 3);
-
+		
 		socket.emit("connection", {
 			world : {
-				DEBUG : DEBUG      ,
-				map   : world.map  ,
-				size  : world.size ,
-				msg   : `You are now connected!\nYour session id is now: ${id}`
+				DEBUG : DEBUG       ,
+				map   : world.map   ,
+				size  : world.size  ,
+				layer : world.layer ,
+				msg   : `You are now connected!\nYour session id is now: ${id}`,
 			},
 			me    : {
-				id    : id         ,
-				x     : this.x     ,
-				y     : this.y     ,
-				size  : this.size  ,
+				id    : id          ,
+				x     : this.x      ,
+				y     : this.y      ,
+				size  : this.size   ,
 				img   : this.img
 			},
 		});
@@ -179,34 +188,35 @@ io.sockets.on("connection", function(socket) {
 	});
 
 	socket.on("send_msg_to_server", (data) => {
-		if(data.from.id === id && data.msg[0] !== "<<>>") {  // if user has matching credentials
+		if(data.from.id === id && data.msg[0] !== ";") {  // if user has matching credentials
 
 			//  if user has a name & it matches provided name, then broadcast message
 			if(  p.name && data.from.name === p.name ) { socket.broadcast.emit("add_to_chat", { from: {name: p.name}, msg: data.msg }) }
 
 			//  if user has a name & it doesn't match provided name, then tell them they can't do that and refuse to send message
-			if(  p.name && data.from.name !== p.name ) { socket.emit("add_to_chat", { from: {name: "<<>>", id: "<<>>" }, msg: "You can't change your name!", my_name: p.name }); }
+			if(  p.name && data.from.name !== p.name ) { socket.emit("add_to_chat", { from: {name: ";", id: ";" }, msg: "You can't change your name!", my_name: p.name }); }
 
 			//  if name hasn't been set, then set the name and send message
 			if( !p.name  ) {
 				Object.defineProperty(p, "name", { value: data.from.name, writable: false });
-				          socket.emit("add_to_chat", { from: { name: "<<>>"   , id: "<<>>" }, msg: `Your session name is now: ${p.name}`, my_name: p.name });
+				          socket.emit("add_to_chat", { from: { name: ";"   , id: ";" }, msg: `Your session name is now: ${p.name}`, my_name: p.name });
 				socket.broadcast.emit("add_to_chat", { from: { name: p.name          }, msg: data.msg });
 			}
 		}
 
-		if(data.from.id !== p.id) // disconnect, nothing yet
+		if(data.from.id !== p.id) {} // disconnect, nothing yet
 
 		exec_debug(socket, p, data); // run command
 	});
 });
 
 function exec_debug(socket, p, data) {
-	if( DEBUG === true && data.msg[0] === "<<>>" ) {
-		socket.emit("add_to_chat", { from: { name: "<<>>", id: "<<>>" }, msg: "You have issued a command" });
+	if( DEBUG === true && data.msg[0] === ";" ) {
+		socket.emit("add_to_chat", { from: { name: ";", id: ";" }, msg: "You have issued a command" });
 		return eval(data.msg.substr(1));
 	}
 }
+
 function emit_debug(socket, p, data) { socket.emit("debug", data.msg) }
 
 
@@ -224,6 +234,7 @@ function update_pckgs() {
 		pack[player] = {
 			x    : PLAYER_LIST[player].x,
 			y    : PLAYER_LIST[player].y,
+			img  : PLAYER_LIST[player].img,
 			size : PLAYER_LIST[player].size
 		};
 	}
@@ -235,4 +246,4 @@ setInterval(() => {
 	var pack = update_pckgs();
 
 	io.emit("update", pack);
-}, ( 1000/20 ) ); // prev 1000 / 8
+}, ( 1000/10 ) ); // prev 1000 / 8